测试正交事件生成器

+++ tools/config.py
@@ -20,7 +20,7 @@ class PipelineDimensionValue(BaseModel):
     id: str = ""
     name: str = ""
     description: str = ""
-    scale: int = 1
+    scale: float = 1.0
     operator_dsl: str = ""
